New tests to replace view_test runtime: 15->0.125s
Previously, the 8 tests in view_test were taking 12 seconds to run. By removing
the dependency on the DecisionTestCase, and the user of the test client, the
run time has been dramatically reduced. NOTE - this includes three failing
tests. One looks like a bug in the code, the other two would round out the
tests but are much easier done with a small code change in views.py
If we could do this on all the tests, the test suite would drop in run time
to under 1s!!