Update add organizations to only use name.
Previously adding an organization required you to specify and email address
and a slug. Now the slug is autogenerated from the name and the email is
auto populated to be the requesting user, so that they see their newly created
organization (they can still add other members or editors later).