Re-applied Sarah's change (https://github.com/aptivate/econsensus/commit/959c95bcfb458bc34084ea2e170c5ea31092e06c - accidentally dropped during conflict resolution on this branch) to
Changed the control for adding a new org from being a sentence at bottom
of screen to being a link at top of screen (orginially done by Sarah in
959c95bc, but accidentally dropped during a conflict resolution on this
branch). Also altered the 'no orgs here' message to refer to new
location of the control, and added some extra detail and formatting to
make the text look more appealing. Tidied up the org list, fixing some
logic that was meant to indicate when the current user is the owner of an
org (but wasn't), and making this into a new section on the line which also
shows the option to delete the org.