Change CustomOrganizationUserLeave to use new template
The template has different text and buttons - all about leaving rather
than deleting.  Also change it so that it works out the user from the
request rather than being passed the id.

And do some refactoring of the CustomOrganizationUserDelete view to make
it easier to reuse in the Leave view.  And refactor the tests to match.