Add References: email header for better threading
particularly for comments.

The References: header has a list of all message-id:s of messages that
have come before it, so it allows threading even if the last message
wasn't received by you (as it wouldn't be if you entered it yourself on
the site).

Before this, if you enter a feedback on the site, then comments to the
feedback won't find the message-id referenced in the In-Reply-To:
header, so the comment email will start a new thread.